src/requirements.txt
scipy
uvicorn
numpy
pandas
scikit-learn
nltk
ipython
matplotlib
seaborn
umap-learn
scipy
numpy
pandas
scikit-learn
nltk